#294 Implement hook
