+++ components/elements/PieChart.tsx
@@ -6,8 +6,7 @@ import { useRef } from 'react';
 import { Chart as ChartJS, ArcElement, Tooltip } from 'chart.js';
 import { Pie } from 'react-chartjs-2';
 import { ChartOptions } from 'chart.js/auto';
-import { useContext } from 'react';
-import { PropsContext } from '../../pages/_app';
+import { useProps } from '../hooks/useProps';
 
 ChartJS.register(ArcElement, Tooltip, ChartDataLabels);
 
@@ -28,7 +27,7 @@ export default function PieChart({
   data: PieChartData | undefined;
   isLoading: boolean;
 }): JSX.Element {
-  const { userSettings } = useContext(PropsContext);
+  const { userSettings } = useProps();
   const chartRef = useRef<ChartJS<'pie'>>(null);
   const sum = data?.data?.reduce((a: number, b: number) => a + b, 0) ?? 0;
